+++ sunpy/net/jsoc/__init__.py
@@ -1,4 +1,2 @@
-from __future__ import absolute_import
-
-from .jsoc import *
-from .attrs import *
+from sunpy.net.jsoc.jsoc import *
+from sunpy.net.jsoc.attrs import *